Migrate tests to Image class
